refactor(backend): move pasword encorder file place
