+++ server/password/password_encoder_test.go
@@ -1,4 +1,4 @@
-package util
+package password
 
 import (
 	"github.com/stretchr/testify/assert"
