+++ server/repository/gorm/repository.go
@@ -1,19 +1,19 @@
 package gorm
 
 import (
-	"github.com/growthers/enshita/server/util"
+	"github.com/growthers/enshita/server/password"
 	"gorm.io/gorm"
 )
 
 type Repository struct {
 	db              *gorm.DB
-	passwordEncoder util.PasswordEncoder
+	passwordEncoder password.Encoder
 }
 
-func NewGormRepository(db *gorm.DB, paswodEncorder util.PasswordEncoder) *Repository {
+func NewGormRepository(db *gorm.DB, passwordEncoder password.Encoder) *Repository {
 	repo := &Repository{
 		db:              db,
-		passwordEncoder: paswodEncorder,
+		passwordEncoder: passwordEncoder,
 	}
 
 	return repo
